feat: handling errors
